requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
